Move package-dependent GPU renderers to their satellite host projects
The base SkiaSharp.Tests.Console must not take a SharpVk (or Vortice)
dependency — it is the host the MAUI device and WASM builds consume
indirectly, and SkiaSharp already ships dedicated satellite projects
(SkiaSharp.Vulkan.Tests.Console, SkiaSharp.Direct3D.Tests.Console) for
exactly these GPU backends, both already built and run by tests-netcore.

- Revert the SharpVk PackageReference + SkiaSharp.Vulkan.SharpVk
  ProjectReference accidentally added to the base Console project.
- Extract the per-cell pipeline (render -> emit ##SKIA-GOLDEN-IMAGE## ->
  compare-or-fail / fail-unseeded) into a reusable VisualMatrixTestsBase
  so every host shares one engine. The shared VisualMatrixTests becomes a
  thin [Theory] over the renderers auto-discovered in the base assembly.
- Move GaneshVulkanRenderer into the Vulkan satellite
  (tests/VulkanTests/Visual/) beside a thin VulkanVisualTests driver that
  runs only the renderers compiled into that assembly via
  RendererCatalog.NamesIn(thisAssembly) -- so raster/GL/Metal are never
  double-run, and a future graphite-vulkan renderer joins automatically.
- Make GpuRenderGate and RendererPixels public (the satellite renderer
  consumes them).
- Update documentation/dev/golden-image-tests.md (hosting/wiring, extend,
  coverage, Graphite seam) for the satellite structure.

Verified on macOS: base Console matrix 15 pass / 0 skip; Vulkan satellite
5 ganesh-vulkan cells skip cleanly (no ICD).
